requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.0
easygui==0.98.3
idna==3.4
imutils==0.5.4
mysql-connector-python==8.2.0
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
p==1.4.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.21.12
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tk==0.1.0
tzdata==2023.3
urllib3==2.1.0
